Fix a bug in zonal_wave_contributions
The initial call to zonal_wave_coeffs inside zonal_wave_contributions erroneously set the waves kwarg to "waves=waves", which kept only the Fourier coefficients for the user-defined wavenumbers. However, the inverse fft needs all the coefficients to reconstruct the correct-sized longitude coordinate. Should consider putting  together a more memory efficient solution such that it's unnecessary to keep the undesired coefficients, but for now, this works.